Add credits screen
The credits screen is accessible by holding A+B and pressing Start on
the title screen.

The palettes and some character data needed to be modified for the
credits screen to work.  Notably, the large "PONG" and "GAME OVER" text
needed to be moved to a higher address space to allow room for a
lower-case font.  Other minor tweaks were to the attribute tables for
the title and game over screens to hide the checkerboard background
pattern with the text.